perf: narrow reconstruct block eob

+++ internal/av1/threading/batch_test.go
@@ -2132,7 +2132,7 @@ func FuzzFrameWorkBatchReconstructBlockCoeff(f *testing.F) {
 			Transform: req.Transform,
 			Quantizer: q,
 			Lossless:  lossless,
-			EOB:       int(req.Block.Result.EOB),
+			EOB:       int16(req.Block.Result.EOB),
 		})
 		if err != nil {
 			return
@@ -2478,7 +2478,7 @@ func testBlockCoeffScratch(t *testing.T, ctx FrameWorkBatch, req FrameWorkBlockC
 		Quantizer:      q,
 		InverseQMatrix: iqMatrix,
 		Lossless:       lossless,
-		EOB:            int(req.Block.Result.EOB),
+		EOB:            int16(req.Block.Result.EOB),
 	}
 	int32Len, int16Len, err := reconstruct.ScratchLen(cfg)
 	if err != nil {
@@ -2506,7 +2506,7 @@ func testReconstructBlockCoeffDirect(t *testing.T, ctx FrameWorkBatch, dst *fram
 		Transform: req.Transform,
 		Quantizer: q,
 		Lossless:  lossless,
-		EOB:       int(req.Block.Result.EOB),
+		EOB:       int16(req.Block.Result.EOB),
 	}
 	int32Scratch, residualScratch := testBlockCoeffScratch(t, ctx, req, plane)
 	dstPlane, _, _, ok := frameWorkFramePlane(dst, plane)
